backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
pillow==11.1.0
numpy==1.24.3
transformers==4.48.0
deep-translator==1.11.4
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
nltk==3.8.1
pycocoevalcap==1.2
rouge==1.0.1